Final Cut

+++ task_3.java
@@ -9,11 +9,11 @@ public class task_3 {
     public static void main(String[] args) {
         Scanner symbol = new Scanner(System.in);
         System.out.print("Enter num1 --> ");
-        float num1 = symbol.nextInt();
+        float num1 = symbol.nextFloat();
         System.out.print("Enter action --> ");
         String action = symbol.next();
         System.out.print("Enter num2 -->  ");
-        float num2 = symbol.nextInt();
+        float num2 = symbol.nextFloat();
         float resultFloat = 0;
         String a = "+";
         String b = "-";
